@@ -15,10 +15,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 0.7.0+1
+version: 0.8.0+1
 
 environment:
-  sdk: '>=2.16.2 <3.0.0'
+  sdk: '^3.2.0'
   flutter: ">=3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
@@ -35,7 +35,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  fluent_ui: 4.8.7
+  fluent_ui: 4.9.2
   flutter_hooks: ^0.20.5
   flag: ^7.0.0
   flutter_svg: ^2.0.10+1
@@ -45,8 +45,8 @@ dependencies:
   web_socket_channel: ^3.0.0
   http: ^1.2.1
   download: ^1.0.0
-  just_audio: ^0.9.29
-  just_audio_windows: ^0.2.1
+  just_audio: ^0.9.41
+  just_audio_windows: ^0.2.2
   path_provider: ^2.0.9
   uuid: ^4.4.0
   tuple: ^2.0.0
@@ -55,6 +55,7 @@ dependencies:
   url_launcher: ^6.1.0
   collection: ^1.16.0
   audioplayers: ^6.0.0
+  twitch_chat: ^1.3.2
 
 dev_dependencies:
   flutter_test:
