@@ -11,12 +11,21 @@ import 'package:streamkit_tts/models/config_model.dart';
 import 'package:streamkit_tts/models/enums/languages_enum.dart';
 import 'package:streamkit_tts/models/enums/tts_source.dart';
 import 'package:streamkit_tts/screens/home/home.dart';
-import 'package:streamkit_tts/screens/trakteer/trakteer.dart';
-import 'package:streamkit_tts/services/chat_to_speech_service.dart';
+import 'package:streamkit_tts/services/composers/app_composer_service.dart';
+import 'package:streamkit_tts/services/composers/composer_service.dart';
+import 'package:streamkit_tts/services/middlewares/bsr_middleware.dart';
+import 'package:streamkit_tts/services/middlewares/forced_language_middleware.dart';
+import 'package:streamkit_tts/services/middlewares/language_middleware.dart';
+import 'package:streamkit_tts/services/middlewares/name_fix_middleware.dart';
+import 'package:streamkit_tts/services/middlewares/pachify_middleware.dart';
+import 'package:streamkit_tts/services/middlewares/read_username_middleware.dart';
+import 'package:streamkit_tts/services/middlewares/remove_emotes_middleware.dart';
+import 'package:streamkit_tts/services/middlewares/remove_urls_middleware.dart';
+import 'package:streamkit_tts/services/middlewares/skip_empty_middleware.dart';
+import 'package:streamkit_tts/services/middlewares/skip_exclamation_middleware.dart';
+import 'package:streamkit_tts/services/middlewares/user_filter_middleware.dart';
 import 'package:streamkit_tts/services/version_check_service.dart';
-import 'package:streamkit_tts/utils/beat_saver_util.dart';
 import 'package:streamkit_tts/utils/external_config_util.dart';
-import 'package:streamkit_tts/utils/language_detection_util.dart';
 import 'package:streamkit_tts/utils/misc_tts_util.dart';
 import 'package:version/version.dart';
 
@@ -32,12 +41,28 @@ void main() async {
     configPath: externalConfigUtil.configPath,
     appPath: externalConfigUtil.appPath,
   );
-  final chatToSpeechService = ChatToSpeechService(
+  final ComposerService composerService = AppComposerService(
     config: config,
-    languageDetectionUtil: LanguageDetection(),
-    externalConfigUtil: externalConfigUtil,
-    miscTtsUtil: MiscTts(),
-    beatSaverUtil: BeatSaverUtil(),
+    middlewares: [
+      // Filter and command handler middlewares
+      UserFilterMiddleware(config: config),
+      BsrMiddleware(config: config),
+      SkipExclamationMiddleware(config: config),
+
+      // Message clean-up middlewares
+      RemoveEmotesMiddleware(config: config),
+      RemoveUrlsMiddleware(config: config),
+
+      ForcedLanguageMiddleware(),
+      PachifyMiddleware(
+        externalConfig: externalConfigUtil,
+        miscTtsUtil: MiscTts(),
+      ),
+      LanguageMiddleware(config: config),
+      SkipEmptyMiddleware(config: config),
+      ReadUsernameMiddleware(config: config),
+      NameFixMiddleware(externalConfig: externalConfigUtil),
+    ],
   );
   final versionCheckService = VersionCheckService();
 
@@ -56,8 +81,9 @@ void main() async {
     MultiProvider(
       providers: [
         ChangeNotifierProvider(create: (_) => config),
-        ChangeNotifierProvider(create: (_) => chatToSpeechService),
+        // ChangeNotifierProvider(create: (_) => chatToSpeechService),
         ChangeNotifierProvider(create: (_) => versionCheckService),
+        Provider(create: (_) => composerService),
       ],
       child: const MyApp(),
     ),
@@ -98,7 +124,7 @@ Future<Config> loadConfigurations(
 
       return Config(chatToSpeechConfiguration: config);
     }
-  } catch (e) {}
+  } catch (_) {}
 
   return Config(
     chatToSpeechConfiguration: ChatToSpeechConfiguration(
@@ -114,12 +140,14 @@ Future<Config> loadConfigurations(
       ttsSource: TtsSource.google,
       filteredUsernames: {},
       isWhitelistingFilter: false,
+      ignoreEmptyMessage: true,
+      ignoreUrls: true,
     ),
   );
 }
 
 class MyApp extends HookWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   // This widget is the root of your application.
   @override
@@ -147,11 +175,6 @@ class MyApp extends HookWidget {
                     icon: SvgPicture.asset("assets/images/twitch_icon.svg"),
                     title: const Text("Twitch"),
                     body: const Home(),
-                  ),
-                  PaneItem(
-                    icon: Image.asset('assets/images/trakteer_icon.png'),
-                    title: const Text("Trakteer"),
-                    body: const Trakteer(),
                   ),
                 ],
               )
@@ -193,7 +216,7 @@ class MyApp extends HookWidget {
 }
 
 class StreamKitTitleBar extends HookWidget {
-  const StreamKitTitleBar({Key? key}) : super(key: key);
+  const StreamKitTitleBar({super.key});
 
   @override
   Widget build(BuildContext context) {
