@@ -9,13 +9,12 @@ import 'package:streamkit_tts/screens/home/widgets/config_groups/tts_config_grou
 import 'package:streamkit_tts/screens/home/widgets/footer/toggle_chat_reader_button.dart';
 import 'package:streamkit_tts/screens/home/widgets/twitch_channel_box.dart';
 import 'package:streamkit_tts/screens/home/widgets/footer/volume_control.dart';
-import 'package:streamkit_tts/services/chat_to_speech_service.dart';
-import 'package:streamkit_tts/services/twitch_chat_service.dart';
+import 'package:streamkit_tts/services/composers/composer_service.dart';
 import 'package:streamkit_tts/services/version_check_service.dart';
 import 'package:url_launcher/url_launcher_string.dart';
 
 class HomeNewVersionWidget extends HookWidget {
-  const HomeNewVersionWidget({Key? key}) : super(key: key);
+  const HomeNewVersionWidget({super.key});
 
   @override
   Widget build(BuildContext context) {
@@ -63,14 +62,14 @@ class HomeNewVersionWidget extends HookWidget {
 }
 
 class Home extends HookWidget {
-  const Home({Key? key}) : super(key: key);
+  const Home({super.key});
 
   @override
   Widget build(BuildContext context) {
     final scrollController = useScrollController();
 
     useEffect(() => _errorStateEffect(context));
-    useEffect(() => _chatToSpeechErrorEffect(context));
+    // useEffect(() => _chatToSpeechErrorEffect(context));
 
     return ScaffoldPage(
       header: const PageHeader(
@@ -130,42 +129,16 @@ class Home extends HookWidget {
   }
 
   Function()? _errorStateEffect(BuildContext context) {
-    final errorStream = context.read<ChatToSpeechService>().errorStream;
-    final subscription = errorStream.listen((error) {
-      switch (error) {
-        case TwitchError.timeout:
-          showDialog(
-            context: context,
-            builder: (context) => ContentDialog(
-              title: const Text("Timeout"),
-              content: const Text(
-                  "Fail to connect to channel. Please make sure the username is correct."),
-              actions: [
-                Button(
-                  child: const Text("Ok"),
-                  onPressed: () {
-                    Navigator.pop(context);
-                  },
-                ),
-              ],
-            ),
-          );
-      }
-    });
-
-    return subscription.cancel;
-  }
-
-  // Currently exclusively used for TikTok TTS error.
-  Function()? _chatToSpeechErrorEffect(BuildContext context) {
-    final errorStream =
-        context.read<ChatToSpeechService>().chatToSpeechErrorMessage;
+    final errorStream = context.read<ComposerService>().getErrorStream();
     final subscription = errorStream.listen((errorMessage) {
+      if (!context.mounted) return;
       showDialog(
         context: context,
         builder: (context) => ContentDialog(
-          title: const Text("TikTok TTS Error"),
-          content: Text(errorMessage),
+          title: const Text("Timeout"),
+          content: Text(
+            errorMessage,
+          ),
           actions: [
             Button(
               child: const Text("Ok"),
