@@ -2,9 +2,7 @@ import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:provider/provider.dart';
 import 'package:streamkit_tts/models/config_model.dart';
-import 'package:streamkit_tts/screens/home/widgets/twitch_channel_selection_dialog.dart';
-import 'package:streamkit_tts/services/chat_to_speech_service.dart';
-import 'package:streamkit_tts/services/twitch_chat_service.dart';
+import 'package:streamkit_tts/screens/home/widgets/dialogs/twitch_channel_selection_dialog.dart';
 
 class TwitchChannelBox extends StatelessWidget {
   const TwitchChannelBox({
@@ -51,28 +49,6 @@ class TwitchChannelBox extends StatelessWidget {
               ...channelInfoWidgets,
               OutlinedButton(
                 onPressed: () {
-                  final service = context.read<ChatToSpeechService>();
-                  if (service.state == TwitchState.loading) {
-                    // Changing channel while Twitch service is still trying to load will cause race condition issues.
-                    // Should be easily fixable IF TWITCH_CHAT_SERVICE ISN'T BLACKBOX TO ME.
-                    showDialog(
-                      context: context,
-                      builder: (context) => ContentDialog(
-                        title: const Text("Sorry..."),
-                        content: const Text(
-                            "Cannot change channel while StreamKit is still trying to connect to a channel. Please wait until connection is either complete or fail."),
-                        actions: [
-                          Button(
-                            child: const Text("Ok"),
-                            onPressed: () {
-                              Navigator.pop(context);
-                            },
-                          ),
-                        ],
-                      ),
-                    );
-                    return;
-                  }
                   showTwitchChannelSelection(context);
                 },
                 child: Text(
@@ -87,6 +63,7 @@ class TwitchChannelBox extends StatelessWidget {
 
   void showTwitchChannelSelection(BuildContext context) {
     showDialog(
-        context: context, builder: (context) => TwitchChannelSelectionDialog());
+        context: context,
+        builder: (context) => const TwitchChannelSelectionDialog());
   }
 }
